Вынес методы судьи в отдельные контроллер и сервис
